Update email-finder extension
- credits fix
- credits fix
- Fix lint and formatting for Raycast publish
- Fix reserved Raycast shortcut on Close action
- Refactor types/utils, fix infinite loading on Reveal Email
- icons saved, name changes, credit balance removed
- remove deleted metadata images
- Merge branch \'contributions/merge-1772562842094\' into raycast/publish-sync
- Pull contributions
