+++ extensions/email-finder/src/hooks/useCompanySearch.ts
@@ -4,26 +4,21 @@ import { searchCompanyByName, CompanySearchResult } from "../backend";
 interface UseCompanySearchResult {
   results: CompanySearchResult[];
   isLoading: boolean;
-  error: string | undefined;
 }
 
 export function useCompanySearch(query: string): UseCompanySearchResult {
   const [results, setResults] = useState<CompanySearchResult[]>([]);
   const [isLoading, setIsLoading] = useState(false);
-  const [error, setError] = useState<string | undefined>(undefined);
 
   useEffect(() => {
-    // * Minimum 2 characters before searching
     if (!query || query.trim().length < 2) {
       setResults([]);
       setIsLoading(false);
-      setError(undefined);
       return;
     }
 
     let cancelled = false;
     setIsLoading(true);
-    setError(undefined);
 
     searchCompanyByName(query)
       .then((data) => {
@@ -32,9 +27,8 @@ export function useCompanySearch(query: string): UseCompanySearchResult {
           setIsLoading(false);
         }
       })
-      .catch((err) => {
+      .catch(() => {
         if (!cancelled) {
-          setError(err instanceof Error ? err.message : "Failed to search");
           setIsLoading(false);
         }
       });
@@ -44,5 +38,5 @@ export function useCompanySearch(query: string): UseCompanySearchResult {
     };
   }, [query]);
 
-  return { results, isLoading, error };
+  return { results, isLoading };
 }
