+++ extensions/email-finder/src/utils.ts
@@ -1,0 +1,3 @@
+export function getErrorMessage(err: unknown): string {
+  return err instanceof Error ? err.message : "Unknown error";
+}
